+++ test/test_scenario.py
@@ -6,7 +6,6 @@ from testix import testixexception
 from testix import hook
 from testix import fake
 from testix import DSL
-from testix.expectations import normal_call
 
 class TestScenario:
     def test_EmptyScenario( self ):
